+++ src/utils/httpServices.js
@@ -8,7 +8,7 @@ const axiosInstance = axios.create({
   baseURL: apiConfig.apiUrl
 });
 
-const getToken = () => {
+const getLocalToken = () => {
   const accessToken = getLocalStorage('accessToken');
   if (accessToken) {
     return `Bearer ${accessToken}`;
@@ -19,7 +19,7 @@ const getToken = () => {
 axiosInstance.interceptors.request.use(
   config => {
     // Do something before request is sent
-    config.headers['Authorization'] = getToken();
+    config.headers['Authorization'] = getLocalToken();
     return config;
   },
   error => {