+++ CHANGELOG.md
@@ -9,11 +9,18 @@
 - Add logout keycloak function
 
 ###### Nov 02, 2021
+
 - Update .env Backend endpoint
 - Apply central logout service
 
 ###### Nov 03, 2021
+
 - Build UI Overview
 
 ###### Nov 08, 2021
+
 - Update Webpack Migration document
+
+###### Nov 24, 2021
+
+- Update authentication endpoint